Rename constant/constant.ts to constant/other.ts
